requirements.txt
asgiref==3.5.2
Authlib==1.0.1
cffi==1.15.1
cryptography==38.0.1
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mock==4.0.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
